+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # ActFramework Change Log
 
 **1.8.8**
+* JSON request resolving - binding String to class issue #802
 * APIDoc - make sample email matches sample firstname and lastname #801
+* e2e - random model data generation #800
 * e2e function: It generated invalid email #799
 * `@JSONField(format = "yyyy-MM")` setting not work #798
 * JSON output: default format for `java.util.Date` missing time part #797